ch06/lambda/ingest-with-python/requirements.txt
deltalake
# The AWS layer doesn't include this unfortunately
pyarrow_hotfix
